Move python perf tests to different job
